+++ tests/fixtures/test_farps_latlon/Missions/Saves/foothold_iraq.lua
@@ -1,0 +1,4 @@
+zonePersistance = {}
+zonePersistance['zones'] = { ['Baghdad']={ ['upgradesUsed']=0,['side']=2,['active']=true,['destroyed']={ },['extraUpgrade']={ },['remainingUnits']={ },['wasBlue']=true,['lat_long']={ ['longitude']=44.361488,['latitude']=33.312805,['altitude']=0 },['firstCaptureByRed']=false,['level']=1,['triggers']={ ['missioncompleted']=0 } }, ['Kirkuk']={ ['upgradesUsed']=0,['side']=1,['active']=true,['destroyed']={ },['extraUpgrade']={ },['remainingUnits']={ },['wasBlue']=false,['lat_long']={ ['longitude']=44.392159,['latitude']=35.468249,['altitude']=0 },['firstCaptureByRed']=true,['level']=1,['triggers']={ ['missioncompleted']=0 } } }
+zonePersistance['zonesDetails'] = { ['Baghdad']={ ['hidden']=false }, ['Kirkuk']={ ['hidden']=false } }
+zonePersistance['playerStats'] = { }